fix(client): rename folder to Comments
